+++ llm_guard/util.py
@@ -6,7 +6,7 @@ import logging
 import re
 import sys
 from functools import lru_cache
-from typing import Any, Literal, NamedTuple
+from typing import Any, Literal, NamedTuple, TextIO
 
 import structlog
 
@@ -25,20 +25,59 @@ EXTERNAL_LOGGERS = {
 CHUNK = NamedTuple("CHUNKS", [("start", int), ("end", int)])
 
 
-def configure_logger(log_level: LOG_LEVELS = "INFO"):
+def configure_logger(
+    log_level: LOG_LEVELS = "INFO", render_json: bool = False, stream: TextIO = sys.stdout
+):
     """
     Configures the logger for the package.
 
     Args:
         log_level: The log level to use for the logger. It should be one of the following strings:
             "DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL".
+        render_json: Whether to render log messages in JSON format. Default is False.
+        stream: The stream to write log messages to. Default is sys.stdout.
     """
     logging.basicConfig(
-        format="[%(asctime)s - %(name)s - %(levelname)s] %(message)s",
+        format="%(message)s",
         level=log_level,
-        stream=sys.stdout,
+        stream=stream,
     )
-    structlog.configure(logger_factory=structlog.stdlib.LoggerFactory())
+
+    log_level_to_int = {
+        "CRITICAL": logging.CRITICAL,
+        "FATAL": logging.FATAL,
+        "ERROR": logging.ERROR,
+        "WARN": logging.WARNING,
+        "WARNING": logging.WARNING,
+        "INFO": logging.INFO,
+        "DEBUG": logging.DEBUG,
+        "NOTSET": logging.NOTSET,
+    }
+
+    processors = [
+        structlog.contextvars.merge_contextvars,
+        structlog.processors.add_log_level,
+        structlog.stdlib.PositionalArgumentsFormatter(),
+        structlog.processors.StackInfoRenderer(),
+        structlog.dev.set_exc_info,
+        structlog.processors.format_exc_info,
+        structlog.processors.UnicodeDecoder(),
+        structlog.processors.TimeStamper(fmt="iso"),
+        structlog.processors.dict_tracebacks,
+        structlog.dev.ConsoleRenderer(),
+    ]
+
+    if render_json:
+        processors = processors[:-1]
+        processors += [structlog.processors.JSONRenderer()]
+
+    structlog.configure(
+        context_class=dict,
+        wrapper_class=structlog.make_filtering_bound_logger(log_level_to_int[log_level]),
+        logger_factory=structlog.PrintLoggerFactory(stream),
+        cache_logger_on_first_use=False,
+        processors=processors,
+    )
     for log_name in EXTERNAL_LOGGERS:
         logging.getLogger(log_name).setLevel(logging.WARNING)