+++ docs/changelog.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `Gibberish` scanner has higher threshold to reduce false positives. In addition, it supports changing `labels` to remove overtriggering when `mild gibberish` is detected.
 - `BanCode` scanner was improved to trigger less false-positives.
+- Improved logging to support JSON format.
 
 ### Removed
 -